+++ jeu_arlecchino/zobrist.hh
@@ -2,6 +2,7 @@
 #include "jeu.hh"
 #include <random>
 
+
 /**********************************************************************************
  *             classe qui sert a représenté une configuration du jeu              *
  *             sous forme d'entier selon le principe du hachage zobrist           *
@@ -16,11 +17,11 @@ class zobrist{
     static const std::string pieces[36];
 
     int indicePiece (const std::string &piece) const;
-    unsigned long long int nombreRandom ();
+    unsigned long int nombreRandom ();
     void buildTableauIndice();
-    unsigned long long int buildKeyZobrist(const int &couleur,const board &plateau,const int &profondeur);
-    void switchKeyZobrist(unsigned long long int &cleZobrist,const board &plateau,const deplacement &coupChoisi,const int &couleur,const int &profondeur);
+    unsigned long int buildKeyZobrist(const int &couleur,const board &plateau,const int &profondeur);
+    void switchKeyZobrist(unsigned long int &cleZobrist,const board &plateau,const deplacement &coupChoisi,const int &couleur,const int &profondeur);
 
     private:
-    std::vector<std::vector<std::vector<std::vector<unsigned long long int>>>> _tableau_indice_zobrist;
+    std::vector<std::vector<std::vector<std::vector<unsigned long int>>>> _tableau_indice_zobrist;
 };
